Fix the class naming to be consistent
