+++ wcfsetup/install/files/lib/system/condition/type/user/AbstractIsNullUserConditionType.class.php
@@ -3,8 +3,6 @@
 namespace wcf\system\condition\type\user;
 
 use wcf\data\DatabaseObjectList;
-use wcf\data\user\User;
-use wcf\data\user\UserList;
 use wcf\system\condition\type\AbstractConditionType;
 use wcf\system\condition\type\IDatabaseObjectListConditionType;
 use wcf\system\condition\type\IMigrateConditionType;
@@ -16,20 +14,15 @@ use wcf\system\form\builder\field\BooleanFormField;
  * @copyright 2001-2025 WoltLab GmbH
  * @license GNU Lesser General Public License <http://opensource.org/licenses/lgpl-license.php>
  * @since 6.3
- *
- * @implements IDatabaseObjectListConditionType<UserList<User>, bool>
- * @implements IObjectConditionType<User, bool>
- * @extends AbstractConditionType<bool>
  */
-abstract class AbstractUserBooleanConditionType extends AbstractConditionType implements IDatabaseObjectListConditionType, IObjectConditionType, IMigrateConditionType
+abstract class AbstractIsNullUserConditionType extends AbstractConditionType implements IDatabaseObjectListConditionType, IObjectConditionType, IMigrateConditionType
 {
     public function __construct(
         public readonly string $identifier,
         public readonly string $columnName,
         public readonly ?string $migrateKeyName = null,
         public readonly ?string $migrateConditionObjectType = null,
-    ) {
-    }
+    ) {}
 
     #[\Override]
     public function getIdentifier(): string
@@ -53,9 +46,9 @@ abstract class AbstractUserBooleanConditionType extends AbstractConditionType im
     public function applyFilter(DatabaseObjectList $objectList): void
     {
         if ($this->filter) {
-            $objectList->getConditionBuilder()->add("{$objectList->getDatabaseTableAlias()}.{$this->columnName} = ?", [1]);
+            $objectList->getConditionBuilder()->add("{$objectList->getDatabaseTableAlias()}.{$this->columnName} IS NOT NULL");
         } else {
-            $objectList->getConditionBuilder()->add("{$objectList->getDatabaseTableAlias()}.{$this->columnName} = ?", [0]);
+            $objectList->getConditionBuilder()->add("{$objectList->getDatabaseTableAlias()}.{$this->columnName} IS NULL");
         }
     }
 
@@ -63,9 +56,9 @@ abstract class AbstractUserBooleanConditionType extends AbstractConditionType im
     public function matches(object $object): bool
     {
         if ($this->filter) {
-            return (bool)$object->{$this->columnName};
+            return $object->{$this->columnName} !== null;
         } else {
-            return !$object->{$this->columnName};
+            return $object->{$this->columnName} === null;
         }
     }
