+++ wcfsetup/install/files/lib/system/condition/type/user/AbstractStringUserConditionType.class.php
@@ -10,8 +10,8 @@ use wcf\system\condition\type\IDatabaseObjectListConditionType;
 use wcf\system\condition\type\IMigrateConditionType;
 use wcf\system\condition\type\IObjectConditionType;
 use wcf\system\form\builder\container\PrefixConditionFormFieldContainer;
-use wcf\system\form\builder\field\IntegerFormField;
 use wcf\system\form\builder\field\SingleSelectionFormField;
+use wcf\system\form\builder\field\TextFormField;
 
 /**
  * @author Olaf Braun
@@ -19,16 +19,17 @@ use wcf\system\form\builder\field\SingleSelectionFormField;
  * @license GNU Lesser General Public License <http://opensource.org/licenses/lgpl-license.php>
  * @since 6.3
  *
- * @phpstan-type Filter = array{condition: string, value: int}
+ * @phpstan-type Filter = array{condition: string, value: string}
  * @implements IDatabaseObjectListConditionType<UserList<User>, Filter>
  * @implements IObjectConditionType<User, Filter>
  * @extends AbstractConditionType<Filter>
  */
-abstract class AbstractUserIntegerConditionType extends AbstractConditionType implements IDatabaseObjectListConditionType, IObjectConditionType, IMigrateConditionType
+abstract class AbstractStringUserConditionType extends AbstractConditionType implements IDatabaseObjectListConditionType, IObjectConditionType, IMigrateConditionType
 {
     public function __construct(
         public readonly string $identifier,
         public readonly string $columnName,
+        public readonly ?string $migrateKeyName = null,
         public readonly ?string $migrateConditionObjectType = null,
     ) {
     }
@@ -38,13 +39,12 @@ abstract class AbstractUserIntegerConditionType extends AbstractConditionType im
     {
         return PrefixConditionFormFieldContainer::create($id)
             ->field(
-                IntegerFormField::create("{$id}Value")
-                    ->minimum(0)
+                TextFormField::create("{$id}Value")
                     ->required()
             )
             ->prefixField(
                 SingleSelectionFormField::create("{$id}Condition")
-                    ->options(\array_combine($this->getConditions(), $this->getConditions()))
+                    ->options($this->getConditions())
                     ->required()
             );
     }
@@ -64,31 +64,47 @@ abstract class AbstractUserIntegerConditionType extends AbstractConditionType im
     #[\Override]
     public function applyFilter(DatabaseObjectList $objectList): void
     {
+        ["condition" => $condition, "value" => $value] = $this->filter;
+        $value = \addcslashes($value, '_%');
+
+        $filter = match ($condition) {
+            "_%" => $value . '%',
+            "%_%" => '%' . $value . '%',
+            "%_" => '%' . $value,
+            default => '',
+        };
+
         $objectList->getConditionBuilder()->add(
-            "{$objectList->getDatabaseTableAlias()}.{$this->columnName} {$this->filter['condition']} ?",
-            [$this->filter['value']]
+            $objectList->getDatabaseTableAlias() . ".{$this->columnName} LIKE ?",
+            [$filter]
         );
     }
 
     #[\Override]
     public function matches(object $object): bool
     {
-        return match ($this->filter['condition']) {
-            '=' => $object->{$this->columnName} == $this->filter['value'],
-            '>' => $object->{$this->columnName} > $this->filter['value'],
-            '<' => $object->{$this->columnName} < $this->filter['value'],
-            '>=' => $object->{$this->columnName} >= $this->filter['value'],
-            '<=' => $object->{$this->columnName} <= $this->filter['value'],
-            default => throw new \InvalidArgumentException("Unknown condition: {$this->filter['condition']}"),
+        ["condition" => $condition, "value" => $value] = $this->filter;
+        $value = \strtolower($value);
+        $objectValue = \strtolower($object->{$this->columnName});
+
+        return match ($condition) {
+            "_%" => \str_starts_with($objectValue, $value),
+            "%_%" => \str_contains($objectValue, $value),
+            "%_" => \str_ends_with($objectValue, $value),
+            default => false,
         };
     }
 
     /**
-     * @return string[]
+     * @return array<string, string>
      */
-    protected function getConditions(): array
+    private function getConditions(): array
     {
-        return ["=", ">", "<", ">=", "<="];
+        return [
+            "_%" => "wcf.condition.startsWith",
+            "%_%" => "wcf.condition.contains",
+            "%_" => "wcf.condition.endsWith",
+        ];
     }
 
     #[\Override]
@@ -100,25 +116,21 @@ abstract class AbstractUserIntegerConditionType extends AbstractConditionType im
     #[\Override]
     public function migrateConditionData(array &$conditionData): array
     {
-        $lessThan = $conditionData['lessThan'] ?? null;
-        $greaterThan = $conditionData['greaterThan'] ?? null;
-        $conditions = [];
-
-        if ($lessThan !== null) {
-            $conditions[] = [
-                'identifier' => $this->getIdentifier(),
-                'value' => ["value" => $lessThan, 'condition' => '<'],
-            ];
-        }
-        if ($greaterThan !== null) {
-            $conditions[] = [
-                'identifier' => $this->getIdentifier(),
-                'value' => ["value" => $greaterThan, 'condition' => '>'],
-            ];
+        $value = $conditionData[$this->migrateKeyName] ?? null;
+        if ($value === null) {
+            return [];
         }
 
-        unset($conditionData['lessThan'], $conditionData['greaterThan']);
+        unset($conditionData[$this->migrateKeyName]);
 
-        return $conditions;
+        return [
+            [
+                'identifier' => $this->identifier,
+                'value' => [
+                    'condition' => "%_%",
+                    'value' => $value,
+                ],
+            ],
+        ];
     }
 }
