+++ wcfsetup/install/files/lib/system/condition/type/user/HasTrophyUserConditionType.class.php
@@ -3,6 +3,9 @@
 namespace wcf\system\condition\type\user;
 
 use wcf\data\DatabaseObjectList;
+use wcf\data\trophy\Trophy;
+use wcf\data\trophy\TrophyList;
+use wcf\data\user\trophy\UserTrophyList;
 use wcf\data\user\User;
 use wcf\data\user\UserList;
 use wcf\system\condition\type\AbstractConditionType;
@@ -10,7 +13,7 @@ use wcf\system\condition\type\IDatabaseObjectListConditionType;
 use wcf\system\condition\type\IMigrateConditionType;
 use wcf\system\condition\type\IObjectConditionType;
 use wcf\system\form\builder\field\SelectFormField;
-use wcf\system\language\LanguageFactory;
+use wcf\system\WCF;
 
 /**
  * @author Olaf Braun
@@ -22,7 +25,7 @@ use wcf\system\language\LanguageFactory;
  * @implements IObjectConditionType<User, string>
  * @extends AbstractConditionType<string>
  */
-final class UserLanguageConditionType extends AbstractConditionType implements IDatabaseObjectListConditionType, IObjectConditionType, IMigrateConditionType
+final class HasTrophyUserConditionType extends AbstractConditionType implements IDatabaseObjectListConditionType, IObjectConditionType, IMigrateConditionType
 {
     #[\Override]
     public function getFormField(string $id): SelectFormField
@@ -30,27 +33,31 @@ final class UserLanguageConditionType extends AbstractConditionType implements I
         // SelectFormField stores its value as a string,
         // so we need to convert it to an integer in the `applyFilter`&`matches` method.
         return SelectFormField::create($id)
-            ->options(LanguageFactory::getInstance()->getLanguages())
+            ->options($this->getTrophies())
             ->required();
     }
 
     #[\Override]
     public function getIdentifier(): string
     {
-        return 'language';
+        return 'hasTrophy';
     }
 
     #[\Override]
     public function getLabel(): string
     {
-        return 'wcf.condition.user.language';
+        return 'wcf.condition.user.hasTrophy';
     }
 
     #[\Override]
     public function applyFilter(DatabaseObjectList $objectList): void
     {
         $objectList->getConditionBuilder()->add(
-            "{$objectList->getDatabaseTableAlias()}.languageID = ?",
+            "{$objectList->getDatabaseTableAlias()}.userID IN (
+                    SELECT userID
+                    FROM   wcf1_user_trophy
+                    WHERE  trophyID = ?
+            )",
             [(int)$this->filter]
         );
     }
@@ -58,25 +65,46 @@ final class UserLanguageConditionType extends AbstractConditionType implements I
     #[\Override]
     public function matches(object $object): bool
     {
-        return (int)$this->filter === $object->languageID;
+        $userTrophies = UserTrophyList::getUserTrophies([$object->userID], false)[$object->userID];
+        $trophyIDs = \array_column($userTrophies, 'trophyID');
+
+        return \in_array((int)$this->filter, $trophyIDs, true);
+    }
+
+    /**
+     * @return Trophy[]
+     */
+    private function getTrophies(): array
+    {
+        $trophyList = new TrophyList();
+        $trophyList->readObjects();
+        $trophies = $trophyList->getObjects();
+
+        $collator = new \Collator(WCF::getLanguage()->getLocale());
+        \uasort(
+            $trophies,
+            static fn (Trophy $a, Trophy $b) => $collator->compare($a->getTitle(), $b->getTitle())
+        );
+
+        return $trophies;
     }
 
     #[\Override]
     public function migrateConditionData(array &$conditionData): array
     {
-        if (!isset($conditionData['languageIDs'])) {
+        if (!isset($conditionData['userTrophyIDs'])) {
             return [];
         }
 
         $result = [];
-        foreach ($conditionData['languageIDs'] as $languageID) {
+        foreach ($conditionData['userTrophyIDs'] as $trophyID) {
             $result[] = [
                 'identifier' => $this->getIdentifier(),
-                'value' => $languageID,
+                'value' => $trophyID,
             ];
         }
 
-        unset($conditionData['languageIDs']);
+        unset($conditionData['userTrophyIDs']);
 
         return $result;
     }
@@ -84,6 +112,6 @@ final class UserLanguageConditionType extends AbstractConditionType implements I
     #[\Override]
     public function canMigrateConditionData(string $objectType): bool
     {
-        return $objectType === 'com.woltlab.wcf.user.languages';
+        return $objectType === 'com.woltlab.wcf.user.userTrophyCondition';
     }
 }
