+++ wcfsetup/install/files/lib/system/condition/type/user/RegistrationDaysUserConditionType.class.php
@@ -13,7 +13,7 @@ use wcf\util\DateUtil;
  * @license GNU Lesser General Public License <http://opensource.org/licenses/lgpl-license.php>
  * @since 6.3
  */
-final class UserRegistrationDaysConditionType extends AbstractUserIntegerConditionType
+final class RegistrationDaysUserConditionType extends AbstractIntegerUserConditionType
 {
     public function __construct()
     {
